Move ajaxDataTable initialization to datatable default script
