+++ app/grandchallenge/core/javascript/datatables.defaults.mjs
@@ -1,7 +1,8 @@
-import $ from "jquery";
-
 import DataTable from "datatables.net-bs4";
+import $ from "jquery";
 import "datatables.net-buttons-bs4";
+
+import { renderVegaChartsObserver } from "../../charts/javascript/render_charts.mjs";
 
 Object.assign(DataTable.defaults, {
     scrollX: true,
@@ -63,7 +64,7 @@ document.addEventListener("DOMContentLoaded", () => {
         if (table.hasAttribute(DT_INITIALIZED_ATTRIBUTE)) {
             continue; // Skip if already initialized
         }
-        const options = {};
+        let options = {};
         for (const attr of table.attributes) {
             if (attr.name.startsWith(DT_ATTRIBUTE_PREFIX)) {
                 // Convert data-dt-foo-bar to fooBar
@@ -77,6 +78,51 @@ document.addEventListener("DOMContentLoaded", () => {
                 options[camelKey] = value;
             }
         }
+
+        if (table.id === "ajaxDataTable") {
+            const defaultSortColumn = JSON.parse(
+                document.getElementById("defaultSortColumn").textContent,
+            );
+            const textAlign = JSON.parse(
+                document.getElementById("textAlign").textContent,
+            );
+            const defaultSortOrder = JSON.parse(
+                document.getElementById("defaultSortOrder").textContent,
+            );
+            renderVegaChartsObserver.observe(
+                document.getElementById("ajaxDataTable"),
+                {
+                    childList: true,
+                    subtree: true,
+                },
+            );
+            const ajaxTableOptions = {
+                order: [[defaultSortColumn, defaultSortOrder]],
+                lengthChange: false,
+                pageLength: 25,
+                serverSide: true,
+                columnDefs: [
+                    ...$.fn.dataTable.defaults.columnDefs,
+                    {
+                        className: `align-middle text-${textAlign}`,
+                        targets: "_all",
+                    },
+                ],
+                ajax: {
+                    url: ".",
+                },
+                ordering: true,
+                drawCallback: settings => {
+                    // trigger htmx process after the page has been updated.
+                    htmx.process("#ajaxDataTable");
+                },
+            };
+            options = {
+                ...options,
+                ...ajaxTableOptions,
+            };
+        }
+
         // Prefer DataTable global, fallback to jQuery if available
         if (typeof window.DataTable !== "undefined") {
             new window.DataTable(table, options);
