+++ app/grandchallenge/datatables/templates/datatables/list_base.html
@@ -11,6 +11,4 @@
     {{ default_sort_column|json_script:"defaultSortColumn" }}
     {{ text_align|json_script:"textAlign" }}
     {{ default_sort_order|json_script:"defaultSortOrder" }}
-
-    <script type="module" src="{% static 'js/datatables/list.mjs' %}"></script>
 {% endblock %}
