+++ app/grandchallenge/datatables/templates/datatables/partials/datatable.html
@@ -1,5 +1,5 @@
 <div class="table-responsive mb-3">
-    <table id="ajaxDataTable" class="table table-hover table-borderless table-sm w-100 no-wrap">
+    <table data-data-table id="ajaxDataTable" class="table table-hover table-borderless table-sm w-100 no-wrap">
         <thead class="thead-light">
         <tr>
             {% for column in columns %}
